pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btcusd-analyzer"
version = "0.1.0"
description = "Neural Network basierte BTC Trendwechsel-Erkennung"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "BTCUSD Analyzer Team"}
]
keywords = ["bitcoin", "trading", "neural-network", "lstm", "transformer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "torch>=2.0.0",
    "PyQt6>=6.5.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "optuna>=3.0.0",
    "python-binance>=1.0.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.2.0",
    "pytorch-forecasting>=1.0.0",
    "einops>=0.6.0",
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "websocket-client>=1.6.0",
    "websockets>=12.0",
    "colorlog>=6.7.0",
    "werkzeug>=3.0.0",
]

[project.optional-dependencies]
backtesting = [
    "vectorbt>=0.26.0",
    "backtrader>=1.9.78",
    "backtesting>=0.3.3",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
btcusd-analyzer = "btcusd_analyzer.main:main"

[project.urls]
Homepage = "https://github.com/btcusd-analyzer/btcusd-analyzer"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=btcusd_analyzer --cov-report=term-missing"
